修复bug

+++ src/main/java/com/gaea/single/bridge/dto/account/ShareRankingRes.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ApiModel("分享排名奖励")
@@ -21,5 +23,5 @@ public class ShareRankingRes {
   private String portraitUrl;
 
   @ApiModelProperty(value = "奖励金额(元)", required = true)
-  private Long money;
+  private BigDecimal money;
 }
